+++ src/components/dashboard/dashboardTabPanes/overview/OverviewMessages.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import {LABELS} from "../../../../helpers/constants/textConstants/labelsAndHeadings";
 import {useAppContext} from "../../../../context/AppContext";
-import {NoDataAvailable} from "../../../minis/NoDataAvailable";
 import {SimpleMessagesList} from "../../../message/SimpleMessagesList";
 
 
@@ -9,16 +8,11 @@ export const OverviewMessages = () => {
 
     const {userMessages, setUserMessages} = useAppContext();
 
-    return (
+    return userMessages && !!userMessages.length && (
         <div className={"sms-dashboard-col--sm"}>
             <div className={"sms-section--light h-100"}>
                 <h2>{LABELS.SECTIONS.DASHBOARD.OVERVIEW.INCOMING_MESSAGES}</h2>
-                {
-                    userMessages && userMessages.length ?
-                        <SimpleMessagesList messagesData={userMessages} setMessagesData={setUserMessages}/>
-                        :
-                        <NoDataAvailable/>
-                }
+                <SimpleMessagesList messagesData={userMessages} setMessagesData={setUserMessages}/>
             </div>
         </div>
     )
