+++ src/components/pages/titles/Title.jsx
@@ -1,5 +1,5 @@
 import React, {useCallback, useEffect, useState} from "react";
-import {HeadingWithBreadCrumbs} from "../headings";
+import {HeadingWithBreadCrumbs} from "../../headings/index.jsx";
 import {useParams} from "react-router-dom";
 import {
     addTitleToTable,
@@ -7,12 +7,12 @@ import {
     handleCollectingTitle,
     removeTitleFromTable,
     userTitleExists
-} from "../../services/serviceFunctions";
-import {ROUTES} from "../../helpers/constants/configConstants";
-import {LABELS} from "../../helpers/constants/textConstants/labelsAndHeadings";
-import {PANES, TEXTS} from "../../helpers/constants/textConstants/texts";
-import {TABLES} from "../../helpers/constants/serviceConstants";
-import {IssuesList} from "../lists/issues/IssuesList";
+} from "../../../services/serviceFunctions.js";
+import {ROUTES} from "../../../helpers/constants/configConstants.jsx";
+import {LABELS} from "../../../helpers/constants/textConstants/labelsAndHeadings.js";
+import {PANES, TEXTS} from "../../../helpers/constants/textConstants/texts.js";
+import {TABLES} from "../../../helpers/constants/serviceConstants.js";
+import {IssuesList} from "../../lists/issues/IssuesList.jsx";
 import {faArrowUpRightFromSquare} from "@fortawesome/pro-regular-svg-icons";
 import {
     faGrid,
@@ -23,27 +23,27 @@ import {
     faPlus,
     faHeart
 } from "@fortawesome/pro-duotone-svg-icons";
-import {getCalculatedYear, getTitleProgressForUser, objectDoesExist} from "../../helpers/functions";
-import {ImageViewerSmall} from "./pagecomponents/ImageViewerSmall";
-import {OverlaySpinner} from "../minis/OverlaySpinner";
-import {useAppContext} from "../../context/AppContext";
-import {getIssuesWithTitleAndPublisherAndGradeValuesByTitleId} from "../../services/issueService";
-import {FunctionButton} from "../minis/FunctionButton";
-import {TitleProgress} from "./TitleProgress";
-import {FormatBadge} from "../minis/FormatBadge";
+import {getCalculatedYear, getTitleProgressForUser, objectDoesExist} from "../../../helpers/functions.jsx";
+import {ImageViewerSmall} from "../pagecomponents/ImageViewerSmall.jsx";
+import {OverlaySpinner} from "../../minis/OverlaySpinner.jsx";
+import {useAppContext} from "../../../context/AppContext.jsx";
+import {getIssuesWithTitleAndPublisherAndGradeValuesByTitleId} from "../../../services/issueService.js";
+import {FunctionButton} from "../../minis/FunctionButton.jsx";
+import {TitleProgress} from "./TitleProgress.jsx";
+import {FormatBadge} from "../../minis/FormatBadge.jsx";
 import {
     addIssueToCollection,
     checkGradingStatus,
     deleteAllGradesByUserAndIssue,
     deleteIssueFromCollectionSimple
-} from "../../services/collectingService";
-import {Message} from "../message/Message";
-import {Icon, editIconDuoTone, infoIconDuoTone, titlesIconDuoTone, valueIconDuoTone} from "../icons";
-import {IconLink} from "../minis/IconLink";
-import {useCollectingStatus} from "../../helpers/customHooks/useCollectingStatus";
-import {SeriekatalogenTitleLink} from "../minis/SeriekatalogenTitleLink";
-import {NoMatch} from "../routes/NoMatch";
-import {MESSAGES} from "../../helpers/constants/textConstants/messages.js";
+} from "../../../services/collectingService.js";
+import {Message} from "../../message/Message.jsx";
+import {Icon, editIconDuoTone, infoIconDuoTone, titlesIconDuoTone, valueIconDuoTone} from "../../icons/index.jsx";
+import {IconLink} from "../../minis/IconLink.jsx";
+import {useCollectingStatus} from "../../../helpers/customHooks/useCollectingStatus.js";
+import {SeriekatalogenTitleLink} from "../../minis/SeriekatalogenTitleLink.jsx";
+import {NoMatch} from "../../routes/NoMatch.jsx";
+import {MESSAGES} from "../../../helpers/constants/textConstants/messages.js";
 
 
 export const Title = () => {
