+++ src/components/routes/MyRoutes.jsx
@@ -8,8 +8,8 @@ import {OverviewPane} from "../dashboard/dashboardTabPanes/overview/OverviewPane
 import {MyTitlesPane} from "../dashboard/dashboardTabPanes/myTitles/MyTitlesPane";
 import {CollectionsPane} from "../dashboard/dashboardTabPanes/collections/CollectionsPane";
 import Profile from "../pages/Profile";
-import {Title} from "../pages/Title";
-import {Titles} from "../pages/Titles";
+import {Title} from "../pages/titles/Title.jsx";
+import {Titles} from "../pages/titles/Titles.jsx";
 import {useAppContext} from "../../context/AppContext";
 import {Admin} from "../pages/admin/Admin";
 import {AdminPublisher} from "../pages/admin/publisher/AdminPublisher";
@@ -27,8 +27,8 @@ import {Publishers} from "../pages/Publishers";
 import {AdminUsers} from "../pages/admin/users/AdminUsers";
 import {Marvelklubben} from "../pages/Marvelklubben";
 import ChangePassword from "../pages/ChangePassword";
-import {User} from "../pages/User";
-import {Users} from "../pages/Users";
+import {User} from "../pages/users/User.jsx";
+import {Users} from "../pages/users/Users.jsx";
 import {AdminMessages} from "../pages/admin/message/AdminMessages";
 import {AdminMessage} from "../pages/admin/message/AdminMessage";
 import {GradeValues} from "../pages/GradeValues";
